requirements.txt
alembic==1.4.1
azure-core==1.10.0
azure-storage-blob==12.7.1
blingfire==0.1.3
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cryptography==3.3.1
cymem==2.0.5
databricks-cli==0.14.1
docker==4.4.1
entrypoints==0.3
Flask==1.1.2
fsspec==0.8.5
gitdb==4.0.5
GitPython==3.1.12
gunicorn==20.0.4
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.4.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
langdetect==1.0.8
lightgbm==3.1.1
Mako==1.1.4
MarkupSafe==1.1.1
mlflow==1.13.1
msrest==0.6.19
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.1
plac==1.1.3
preshed==3.0.5
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.14.0
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
smmap==3.0.4
spacy==2.3.5
SQLAlchemy==1.3.22
sqlparse==0.4.1
srsly==1.0.5
tabulate==0.8.7
thinc==7.4.5
threadpoolctl==2.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
wasabi==0.8.1
websocket-client==0.57.0
Werkzeug==1.0.1
xgboost==1.3.3
zipp==3.4.0
